* Fix error caused by separators (, or ., depending on your locale) messing up values > 1000. * Change use of int for Qualities to Long.

+++ src/org/gmc/ssseditor/qualities/Companion.java
@@ -47,7 +47,7 @@ public class Companion extends QualityItem
 	 */
 	public Companion(Map<String, Object> quality)
 	{
-		super(JSONUtil.getString(quality, "Name"), JSONUtil.getInteger(quality, "Id"), JSONUtil.getString(quality, "AssignToSlot", "Name"), Companion.categoryName);
+		super(JSONUtil.getString(quality, "Name"), JSONUtil.getLong(quality, "Id"), JSONUtil.getString(quality, "AssignToSlot", "Name"), Companion.categoryName);
 		// Other nodes of interest:
 		//   AvailableAt: Text, e.g. "Sometimes available for hire in London."
 		//   Enhancements: Qualities enhanced by this Companion, e.g. Mirrors will be
